backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.0
requests==2.31.0
python-dotenv==1.0.1
supabase==2.3.0
apscheduler==3.10.4
python-jose==3.3.0
